much more stable, should be ready to bundle
