+++ py/PIEdataVARS.py
@@ -1,5 +1,4 @@
 from py import PIEdata
-from py import PIEdata_new
 
 # Make some variables
 contactcatdict = {}
@@ -31,42 +30,52 @@ appointmentstatdict.append('Cancelled')
 list_o_data = []
 
 # CONTACTS
-contacts_new = PIEdata_new.PIEdata('Contacts - Long', 'https://tcciub.pie.iu.edu/Api/Contacts?page=0&pageLimit=100000&searchTerms=')
-contacts_new.categoryDict = contactcatdict
-list_o_data.append(contacts_new)
+contacts_long = PIEdata.PIEdata('Contacts - Long', 'https://tcciub.pie.iu.edu/Api/Contacts?page=0&pageLimit=100000&searchTerms=')
+contacts_long.categoryDict = contactcatdict
+list_o_data.append(contacts_long)
 
 # CONTACTS - mini (lil cutie goobers)
-contacts_fast = PIEdata_new.PIEdata('Contacts', 'https://tcciub.pie.iu.edu/Api/Contacts?page=0&pageLimit=100000&searchTerms=')
-contacts_fast.append = '&mini=true'
-contacts_fast.categoryDict = contactcatdict
-list_o_data.append(contacts_fast)
+contacts = PIEdata.PIEdata('Contacts', 'https://tcciub.pie.iu.edu/Api/Contacts?page=0&pageLimit=100000&searchTerms=')
+contacts.append = '&mini=true'
+contacts.categoryDict = contactcatdict
+list_o_data.append(contacts)
 
 # GOLDSTARS
-goldstars = PIEdata_new.PIEdata('Gold Stars', 'https://tcciub.pie.iu.edu/Api/GoldStars?page=0&pageLimit=100000&searchTerms=')
+goldstars = PIEdata.PIEdata('Gold Stars', 'https://tcciub.pie.iu.edu/Api/GoldStars?page=0&pageLimit=100000&searchTerms=')
 goldstars.createdbyPost = True
 goldstars.statusDict=statusdict
 list_o_data.append(goldstars)
 
 # PDIs
-pdis = PIEdata_new.PIEdata('PDIs', 'https://tcciub.pie.iu.edu/Api/ProfessionalDevelopmentIssues?page=0&pageLimit=100000&searchTerms=')
+pdis = PIEdata.PIEdata('PDIs', 'https://tcciub.pie.iu.edu/Api/ProfessionalDevelopmentIssues?page=0&pageLimit=100000&searchTerms=')
 pdis.createdbyPost = True
 pdis.statusDict = statusdict
 list_o_data.append(pdis)
 
 # LOCATIONS
-locations = PIEdata_new.PIEdata('User Locations', 'https://tcciub.pie.iu.edu/Api/UserLocations?page=0&pageLimit=1000000&onlyLocation=true&searchTerms=')
+locations = PIEdata.PIEdata('User Locations', 'https://tcciub.pie.iu.edu/Api/UserLocations?page=0&pageLimit=1000000&onlyLocation=true&searchTerms=')
 locations.locationPost = True
 locations.createSwitch = True
 list_o_data.append(locations)
 
+# Labs
+labs = PIEdata.PIEdata('Active Labs', 'https://pie.iu.edu/Api/Locations?')
+labs.allowDates = False
+# list_o_data.append(labs)
+
+# Inventory Labs
+inv_labs = PIEdata.PIEdata('Active Inventory Labs', 'https://pie.iu.edu/Api/InventoryLocations?page=0&pageLimit=1010')
+inv_labs.allowDates = False
+# list_o_data.append(inv_labs)
+
 # SHIFTS
-shifts = PIEdata_new.PIEdata('Shifts', 'https://tcciub.pie.iu.edu/Api/Shifts?page=0&pageLimit=100000&weekView=false')
+shifts = PIEdata.PIEdata('Shifts', 'https://tcciub.pie.iu.edu/Api/Shifts?page=0&pageLimit=100000&weekView=false')
 shifts.append = '&minimal=true'
 shifts.locationPost = True
 list_o_data.append(shifts)
 
 # INVENTORY REPORTS
-invreports = PIEdata_new.PIEdata('Inventory Reports', 'https://tcciub.pie.iu.edu/Api/InventoryReports?')
+invreports = PIEdata.PIEdata('Inventory Reports', 'https://tcciub.pie.iu.edu/Api/InventoryReports?')
 invreports.invbool = True
 invreports.chuncks = 15
 invreports.createdbyPost = True
@@ -75,47 +84,47 @@ invreports.append = '&onlyCurrent=false&pageLimit=null'
 list_o_data.append(invreports)
 
 # APPOINTMENTS
-appointments = PIEdata_new.PIEdata('Appointments', 'https://tcciub.pie.iu.edu/Api/Appointments?page=0&pageLimit=10000000')
+appointments = PIEdata.PIEdata('Appointments', 'https://tcciub.pie.iu.edu/Api/Appointments?page=0&pageLimit=10000000')
 appointments.statusDict = appointmentstatdict
 appointments.append = '&groupByEmployee=false&instanceId=1'
 appointments.assignedToPost = True
 list_o_data.append(appointments)
 
 # ACTIVE USERS
-activerusers = PIEdata_new.PIEdata('Active Users', 'https://tcciub.pie.iu.edu/Api/Users?page=0&pageLimit=1000&searchTerms=&active=true&whitelistInclusiveMaskNames=Employee&fromUsersView=true&maskId=5')
+activerusers = PIEdata.PIEdata('Active Users', 'https://tcciub.pie.iu.edu/Api/Users?page=0&pageLimit=1000&searchTerms=&active=true&whitelistInclusiveMaskNames=Employee&fromUsersView=true&maskId=5')
 activerusers.allowDates = False
 list_o_data.append(activerusers)
 
 # GRAVEYARD HEADCOUNTS
-graveyardheads = PIEdata_new.PIEdata('Graveyard HeadCounts', 'https://tcciub.pie.iu.edu/Api/LocationEvaluations/Reports?page=0&pageLimit=10000000')
+graveyardheads = PIEdata.PIEdata('Graveyard HeadCounts', 'https://tcciub.pie.iu.edu/Api/LocationEvaluations/Reports?page=0&pageLimit=10000000')
 graveyardheads.append = '&formId=176&minimal=true'
 list_o_data.append(graveyardheads)
 
 # FORM KEY
-formkey = PIEdata_new.PIEdata('Form Key', 'https://tcciub.pie.iu.edu/Api/LocationEvaluationForms')
+formkey = PIEdata.PIEdata('Form Key', 'https://tcciub.pie.iu.edu/Api/LocationEvaluationForms')
 
 # SUB PLEAS
-subpleas = PIEdata_new.PIEdata('Sub Plea', 'https://tcciub.pie.iu.edu/Api/SubPleas?page=0&pageLimit=1000000')
+subpleas = PIEdata.PIEdata('Sub Plea', 'https://tcciub.pie.iu.edu/Api/SubPleas?page=0&pageLimit=1000000')
 list_o_data.append(subpleas)
 
 # SCHEDULES
-schedules = PIEdata_new.PIEdata('Semester Schedules', 'https://tcciub.pie.iu.edu/Api/Schedules?page=0&pageLimit=101')
+schedules = PIEdata.PIEdata('Semester Schedules', 'https://tcciub.pie.iu.edu/Api/Schedules?page=0&pageLimit=101')
 schedules.allowDates = False
 list_o_data.append(schedules)
 
 # ASSIGNED BADGES
-assignedbadges = PIEdata_new.PIEdata('Assigned Badges', 'https://tcciub.pie.iu.edu/Api/AssignedBadges?page=0&pageLimit=1000000')
+assignedbadges = PIEdata.PIEdata('Assigned Badges', 'https://tcciub.pie.iu.edu/Api/AssignedBadges?page=0&pageLimit=1000000')
 assignedbadges.createdbyPost = True
 assignedbadges.assignedToPost = True
 list_o_data.append(assignedbadges)
 
 # EMPLOYEE MEETINGS
-employeeMeetings = PIEdata_new.PIEdata('Employee Meetings', 'https://tcciub.pie.iu.edu/Api/EmployeeMeetings?page=0&pageLimit=10000000')
+employeeMeetings = PIEdata.PIEdata('Employee Meetings', 'https://tcciub.pie.iu.edu/Api/EmployeeMeetings?page=0&pageLimit=10000000')
 employeeMeetings.employeeswitch = True
 list_o_data.append(employeeMeetings)
 
 # ACCOUNT CHECKS
-accountchecks = PIEdata_new.PIEdata('Account Checks', 'https://tcciub.pie.iu.edu/Api/AccountChecks?page=0&pageLimit=1000000')
+accountchecks = PIEdata.PIEdata('Account Checks', 'https://tcciub.pie.iu.edu/Api/AccountChecks?page=0&pageLimit=1000000')
 accountchecks.createdbyPost = True
 accountchecks.startPost = True
 accountchecks.endPost = True
@@ -123,17 +132,17 @@ accountchecks.locationPost = True
 list_o_data.append(accountchecks)
 
 # ATTENDANCE ISSUES
-attendance_issues = PIEdata_new.PIEdata('Attendance Issues', 'https://tcciub.pie.iu.edu/Api/AttendanceIssues?page=0&pageLimit=1000000')
+attendance_issues = PIEdata.PIEdata('Attendance Issues', 'https://tcciub.pie.iu.edu/Api/AttendanceIssues?page=0&pageLimit=1000000')
 attendance_issues.assignedToPost = True
 list_o_data.append(attendance_issues)
 
 # CHAT MESSAGES
-chat_messages = PIEdata_new.PIEdata('Chat Messages', 'https://tcciub.pie.iu.edu/Api/ChatMessages?page=0&pageLimit=1000000')
+chat_messages = PIEdata.PIEdata('Chat Messages', 'https://tcciub.pie.iu.edu/Api/ChatMessages?page=0&pageLimit=1000000')
 chat_messages.createSwitch = True
 list_o_data.append(chat_messages)
 
 # INCIDENT REPORTS
-incident_reports = PIEdata_new.PIEdata('Incident Reports', 'https://tcciub.pie.iu.edu/Api/IncidentReports?page=0&pageLimit=1000000')
+incident_reports = PIEdata.PIEdata('Incident Reports', 'https://tcciub.pie.iu.edu/Api/IncidentReports?page=0&pageLimit=1000000')
 incident_reports.createdbyPost = True
 incident_reports.startPost = True
 incident_reports.endPost = True
@@ -143,10 +152,10 @@ list_o_data.append(incident_reports)
 def buildalldatathings(userdict, labdict, invlabs):
 
     # buff contacts
-    contacts_new.createdbyDict = userdict
-    contacts_new.locationDict = labdict
-    contacts_fast.createdbyDict = userdict
-    contacts_fast.locationDict = labdict
+    contacts_long.createdbyDict = userdict
+    contacts_long.locationDict = labdict
+    contacts.createdbyDict = userdict
+    contacts.locationDict = labdict
 
     # buff goldstars
     goldstars.createdbyDict = userdict
